sync feed just for old users and just once
