+++ Source/Service/BotMigrationCoordinator.swift
@@ -82,31 +82,9 @@ class BotMigrationCoordinator: BotMigrationDelegate {
     }
     
     @objc func onDone(notification: Notification) {
-        let userDefaults = UserDefaults.standard
-        let didMigrateIdentityToScuttlego = userDefaults.bool(forKey: UserDefaults.didMigrateIdentityToScuttlego)
-        let enableDoneButton = { [botMigrationController] in
-            _ = Task.detached(priority: .high) { @MainActor in
-                do {
-                    try await Bots.current.syncLoggedIdentity()
-                    userDefaults.setValue(true, forKey: UserDefaults.didMigrateIdentityToScuttlego)
-                } catch {
-                    CrashReporting.shared.reportIfNeeded(error: error)
-                }
-                botMigrationController.isDone = true
-                Analytics.shared.trackDidFinishBotMigration()
-            }
-        }
-        if didMigrateIdentityToScuttlego {
-            enableDoneButton()
-        } else {
-            Task.detached(priority: .high) {
-                do {
-                    try await Bots.current.syncLoggedIdentity()
-                } catch {
-                    CrashReporting.shared.reportIfNeeded(error: error)
-                }
-                enableDoneButton()
-            }
+        Task.detached(priority: .high) { @MainActor [botMigrationController] in
+            botMigrationController.isDone = true
+            Analytics.shared.trackDidFinishBotMigration()
         }
     }
 }
